refactor: api 모듈에 도메인 객체 연결
- remote data를 도메인 객체로 변환 후 반환
